split galley API
